fix: add regions fetch to rulesprovider
